requirements.txt
dnspython
openpyxl
